functional help

+++ commands/clear.js
@@ -1,4 +1,5 @@
-const { SlashCommandBuilder } = require('@discordjs/builders')
+const { MessageActionRow, MessageButton, Permissions } = require('discord.js');
+const { SlashCommandBuilder } = require('@discordjs/builders');
 
 
 module.exports = {
@@ -13,10 +14,16 @@ module.exports = {
   
   async execute(interaction) {
     const amount = interaction.options.getInteger('amount')
-    await interaction.channel.bulkDelete(amount)
-      .then(console.log(`Cleared ${amount} messages`))
-      .catch(console.error);
-    await interaction.reply({ content: `Cleared \`${amount}\` messages`, ephemeral: true })
+    
+    if (interaction.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) {
+      interaction.channel.bulkDelete(amount)
+        .then(console.log(`Cleared ${amount} messages`))
+        .catch(console.error);
+      await interaction.reply({ content: `Cleared \`${amount}\` messages`, ephemeral: true })
+    } else {
+      await interaction.reply({ content: `You cannot use the clear command`, ephemeral: false })
+      console.log('BUT FAILED MISERABLY HAHAHAHAHAHHAHAHWAIHUFAIUFAIWF')
+    }
 	}
 }
 
@@ -24,5 +31,5 @@ module.exports.help = {
   name: module.exports.data.name,
   description: module.exports.data.description,
   arguments: "<amount [int 1~100]>",
-  usage: '`/' + module.exports.data.name + '<amount>`'
+  usage: '`/' + module.exports.data.name + ' <amount>`'
 }
